requirements.txt
gTTS
pyaudio
playsound==1.2.2
cx-Freeze
deep-translator
SpeechRecognition
google-transliteration-api
flask
